Refactoring super-rectangle to EnsoGL
