+++ lib/rust/ensogl/core/src/display/shape.rs
@@ -9,9 +9,9 @@ pub mod compound;
 pub mod constants;
 pub mod primitive;
 
+pub use compound::rectangle::Rectangle;
 pub use constants::*;
 pub use primitive::*;
-
 
 
 // We have two Shape and two ShapeOps traits. This one takes precedence.
